Remove plank v1
This change removes plank v1, which has been deprecated since October
2020 and was originally slated for removal in February 2021.

Users are advised to instead use the prow-controller-manager.
