+++ prow/BUILD.bazel
@@ -32,7 +32,6 @@ prow_push(
             "mkpj",
             "mkpod",
             "peribolos",
-            "plank",
             "sidecar",
             "sinker",
             "status-reconciler",
@@ -115,7 +114,6 @@ filegroup(
         "//prow/cmd/phaino:all-srcs",
         "//prow/cmd/phony:all-srcs",
         "//prow/cmd/pipeline:all-srcs",
-        "//prow/cmd/plank:all-srcs",
         "//prow/cmd/prow-controller-manager:all-srcs",
         "//prow/cmd/sidecar:all-srcs",
         "//prow/cmd/sinker:all-srcs",
