--- prow/cmd/plank/BUILD.bazel
@@ -1,61 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
-load("//prow:def.bzl", "prow_image")
-
-NAME = "plank"
-
-prow_image(
-    name = "image",
-    base = "@alpine-base//image",
-    component = NAME,
-    visibility = ["//visibility:public"],
-)
-
-go_binary(
-    name = NAME,
-    embed = [":go_default_library"],
-    pure = "on",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["main.go"],
-    importpath = "k8s.io/test-infra/prow/cmd/plank",
-    deps = [
-        "//pkg/flagutil:go_default_library",
-        "//prow/config:go_default_library",
-        "//prow/flagutil:go_default_library",
-        "//prow/interrupts:go_default_library",
-        "//prow/logrusutil:go_default_library",
-        "//prow/metrics:go_default_library",
-        "//prow/pjutil:go_default_library",
-        "//prow/plank:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@io_k8s_apimachinery//pkg/labels:go_default_library",
-        "@io_k8s_sigs_controller_runtime//pkg/manager:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["main_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//prow/flagutil:go_default_library",
-        "@io_k8s_apimachinery//pkg/util/sets:go_default_library",
-    ],
-)
