--- prow/cmd/plank/OWNERS
@@ -1,4 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-labels:
- - area/prow/plank
