+++ prow/def.bzl
@@ -100,18 +100,12 @@ def target(cmd):
 #     "gcr.io/k8s-prow/hook:20180203-deadbeef": "//prow/cmd/hook:image",
 #     "gcr.io/k8s-prow/hook:latest": "//prow/cmd/hook:image",
 #     "gcr.io/k8s-prow/hook:latest-fejta": "//prow/cmd/hook:image",
-#     "gcr.io/k8s-prow/plank:20180203-deadbeef": "//prow/cmd/plank:image",
-#     "gcr.io/k8s-prow/plank:latest": "//prow/cmd/plank:image",
-#     "gcr.io/k8s-prow/plank:latest-fejta": "//prow/cmd/plank:image",
 #     "gcr.io/k8s-prow/ghproxy:20180203-deadbeef": "//ghproxy:image",
 #     "gcr.io/k8s-prow/ghproxy:latest": "//ghproxy:image",
 #     "gcr.io/k8s-prow/ghproxy:latest-fejta": "//ghproxy:image",
 #     "gcr.io/k8s-prow-edge/hook:20180203-deadbeef": "//prow/cmd/hook:image",
 #     "gcr.io/k8s-prow-edge/hook:latest": "//prow/cmd/hook:image",
 #     "gcr.io/k8s-prow-edge/hook:latest-fejta": "//prow/cmd/hook:image",
-#     "gcr.io/k8s-prow-edge/plank:20180203-deadbeef": "//prow/cmd/plank:image",
-#     "gcr.io/k8s-prow-edge/plank:latest": "//prow/cmd/plank:image",
-#     "gcr.io/k8s-prow-edge/plank:latest-fejta": "//prow/cmd/plank:image",
 #     "gcr.io/k8s-prow-edge/ghproxy:20180203-deadbeef": "//ghproxy:image",
 #     "gcr.io/k8s-prow-edge/ghproxy:latest": "//ghproxy:image",
 #     "gcr.io/k8s-prow-edge/ghproxy:latest-fejta": "//ghproxy:image",
