+++ prow/plank/BUILD.bazel
@@ -20,7 +20,6 @@ go_test(
         "//prow/kube:go_default_library",
         "//prow/pjutil:go_default_library",
         "@com_github_go_test_deep//:go_default_library",
-        "@com_github_mohae_deepcopy//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
@@ -49,7 +48,6 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = [
-        "controller.go",
         "error.go",
         "reconciler.go",
     ],
